// src/types/Synergetic/iSynLuHouse.ts
type iSynLuHouse ={
  Campus: string;
  Code: string;
  Description: string;
  HeadOfHouse: string;
  HeadOfHouseID: number | null;
  HouseCaptain: string;
  ActiveFlag: boolean;
  SortOrder: number;
  ModifiedDate: Date | string | null;
  ModifiedUser: string;
  SetCentrallyFlag: boolean | null;
}

export default iSynLuHouse;